--- game-theory-toolkit/__init__.py
@@ -1,2 +1,0 @@
-from tools.player import Player
-from tools.game import StrategicGame
