+++ game_theory_toolkit/__init__.py
@@ -1,0 +1,2 @@
+from .player import Player
+from .game import StrategicGame
